added functionality to display celebratory name dinamically
