+++ Person-Go/Person-Go/Views/LaunchListView.swift
@@ -11,7 +11,6 @@ struct Friend1: Identifiable {
 struct LaunchListView: View {
     @StateObject private var locationManager = LocationManager()
     @State private var friends: [Friend1] = []
-    @Environment(\.colorScheme) var colorScheme
     @State private var selectedFriends: Set<UUID> = []
     @Binding var selectedTab: String
     @ObservedObject var selectedFriendsStore: SelectedFriends
@@ -31,7 +30,7 @@ struct LaunchListView: View {
                 Divider()
                     .frame(height: 2)
                 
-                List(friends) { friend in
+                List(friends, id: \.id) { friend in
                     FriendRow(friend: friend, isSelected: self.selectedFriends.contains(friend.id))
                         .onTapGesture {
                             if self.selectedFriends.contains(friend.id) {
@@ -93,16 +92,14 @@ struct LaunchListView: View {
             friends.removeAll()
 
             // For each friend in the list, calculate the distance from the user's current location
-            if let userLocation = locationManager.currentLocation {
-                for friend in friendsList {
-                    if let latitude = friend.latitude, let longitude = friend.longitude {
-                        let friendLocation = CLLocation(latitude: latitude, longitude: longitude)
-                        let distance = userLocation.distance(from: friendLocation) / 1000.0 // Convert to kilometers
+            for friend in friendsList {
+                if let latitude = friend.latitude, let longitude = friend.longitude {
+                    let friendLocation = CLLocation(latitude: latitude, longitude: longitude)
+                    let distance = locationManager.calculateDistance(from: locationManager.currentLocation!, to: friendLocation)
 
-                        // Create a new Friend1 object with the calculated distance and add it to the friends array
-                        let newFriend = Friend1(name: friend.username ?? "", distance: distance, avatar: friend.avatarUrl ?? "sample")
-                        friends.append(newFriend)
-                    }
+                    // Create a new Friend1 object with the calculated distance and add it to the friends array
+                    let newFriend = Friend1(name: friend.username ?? "", distance: distance, avatar: friend.avatarUrl ?? "sample")
+                    friends.append(newFriend)
                 }
             }
         } catch {
